+++ include/portable/console.h
