+++ include/portable/mem.h
