+++ include/x86/gdt.h
@@ -7,7 +7,7 @@
  * 
  * 
  */
-#include "types.h"
+#include <portable/types.h>
 // Each define here is for a specific flag in the descriptor.
 // Refer to the intel documentation for a description of what each one does.
