+++ include/x86/idt.h
@@ -1,4 +1,4 @@
-#include "types.h"
+#include <portable/types.h>
 
 #define IDT_SIZE 256
 /// DPL
