+++ include/x86/mouse.h
@@ -1,4 +1,4 @@
-#include "types.h"
+#include <portable/types.h>
 
 void mouse_wait(byte type);
 void mouse_write(byte data);
