+++ include/x86/ports.h
@@ -1,0 +1,2 @@
+int read_port(int num);
+void write_port(int num, int data);
