+++ include/x86/x86.h
@@ -1,4 +1,4 @@
-#include "types.h"
+#include <portable/types.h>
 
 extern int read_port(int num);
 extern void write_port(int num, int data);
