+++ x86/idt.c
@@ -7,9 +7,9 @@
  * 
  * 
  */
-#include "idt.h"
-#include "isr.h"
-#include "x86.h"
+#include <x86/idt.h>
+#include <x86/isr.h>
+#include <x86/x86.h>
 
 /// таблица дескриптеров прерываний
 idtGateDescriptor idt[IDT_SIZE];
