+++ x86/mem_check.asm
